FCN-238 refactor drawer
